Actually describe vague "checkChunk" methods.
These are quite vague, and the profiler call is all we really have to go off of.

From looking around the codebase, these methods validate the entity's chunkX/Y/Z positions are correct. If the entity's last and current chunk pos are not the same, move the entity to it's new chunk (which also changes the chunkX/Y/Z values to the correct ones).

Oh and some synthetic method params while I was at it.