TextFieldWidget tweaks
The field focused clashes with superclass field, and is not accurate. It
actually controls whether the background is drawn. The getter and setter
methods are also not correct, as the field does not just control the
border.

The setSelected method does not set whether the field is selected, but
focused. It only exists to expose the protected AbstractButtonWidget.setFocused
method. It's named setTextFieldFocused not setFocused because that would
clash with the superclass name.