`BlockStateDefinitionProvider` -> `ModelProvider`
`BlockStateDefinitionProvider` is inaccurate, as it also handles all the models in the game including item models.

Also moved everything from `data/client/model` down to just `data/client` as they're *all* model related so there's no point in having it be separate.

Should probably backport this to 1.18, and maybe even 1.17.1 if #2904 is merged. Also recommend the rename is reflected in the Fabric API implementation also.