Rename Structure and StructureType (#3183)
* Rename Structure and StructureType
Structures really need a revisit. In my opinion the net.minecraft.structure package should be gone and
everything should be moved to net.minecraft.world.gen.structure. That'd be a very big change, though.
This commit basically reverts the entirety of #3079 and contains the changes necessary to keep consistency
(but not other new changes). I could make it bigger, but this needs an ASAP merge to stop the bad names from
going to the release.

Summary of the renames:

- `structure.Structure` became `structure.StructureTemplate`. See `/place template`, datafixes, etc.
- `structure.StructureType` became `world.gen.structure.StructureType`.
- `world.gen.structure.StructureType` became `world.gen.structure.Structure`.
- `world.gen.structure.StructureTypes` became `world.gen.structure.Structures`.
- Other names referencing the renamed classes by their old name
- And missed "feature" renames (Fixes #3179)

* Rename StructureManager

* merging gone wrong

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>