Data package restructure (#3465)
With the introduction of sub-packs included in the base game, there are several classes for the 1.20 datapack and several for the vanilla datapack which would benefit from being renamed and put into packages to make this clearer.

Classes which are only used for the vanilla datapack have been put into a `vanilla` subpackage, and ones for the 1.20 datapack have been put into a `onetwenty` subpackage.

Furthermore, classes which are only used for the vanilla datapack have been prefixed with `Vanilla`.

The above changes should not be considered impactful as these classes are strictly for generating the vanilla and 1.20 datapacks and are incredibly unlikely to be referenced by modders.

However, tag providers which were named with an `Abstract` prefix have had this removed. This makes them match with other providers. This is a somewhat impactful refactor.