Improve methods related to closing handled screens (#3505)
* ChestBlockEntity.onInvOpenOrClose -> onViewerCountUpdate

The new name is more accurate - the method takes in the old and new
viewer counts as parameters.

* ScreenHandler.close -> onClosed

Fixes #3503. The method doesn't close anything, it's only called when the screen handler gets closed.

* PlayerEntity.closeScreenHandler -> onHandledScreenClosed

Fixes #2612. The method closes nothing, it just runs the closing tasks
*after* the screen has closed. The current name can easily be confused
with closeHandledScreen(), which does actually close the screen.

* Document PlayerEntity.closeHandledScreen