Fix FoodComponent constructor having outdated names (#3908)
For some reason, likely due to its use of generics in its record components, the FoodComponent record still uses a separate constructor which is visible using some decompilers, and had wrong and outdated mappings for its arguments which weren't visible in enigma. This manually fixes the arguments to match the exposed record component names.